+++ projects/16-open-qf/src/site/components/user/info/index.jsx
@@ -1,9 +1,8 @@
 import { USER_INFO, USER_POWER } from "@/fixtures/user";
-import { cn } from "@/utils";
-import { PieChart } from "react-minimal-pie-chart";
-import Avatar from "../accountSelector/avatar";
-import Copy from "../copy";
-import IdentityOrAddr from "./identityOrAddr";
+import Avatar from "../../accountSelector/avatar";
+import Copy from "../../copy";
+import IdentityOrAddr from "../identityOrAddr";
+import QFpowerPie from "./qfpowerPie";
 
 export default function UserInfo() {
   return (
@@ -49,43 +48,6 @@ function Item({ title, children }) {
     <div className="space-y-1">
       <div className="text16semibold text-text-primary">{title}</div>
       <div>{children}</div>
-    </div>
-  );
-}
-
-function QFpowerPie({ className = "", percentage = 0 }) {
-  const data = [
-    {
-      value: 100 - percentage,
-      color: "var(--fill-bg-quaternary)",
-    },
-    {
-      value: percentage,
-      color: "var(--fill-bg-brand-secondary)",
-    },
-  ];
-
-  return (
-    <div className={cn("relative", className)}>
-      <div className="w-full overflow-hidden">
-        <PieChart
-          // mirrored
-          className={"scale-x-[-1]"}
-          center={[50, 50]}
-          rounded
-          lineWidth={15}
-          data={data}
-          startAngle={-180}
-          lengthAngle={180}
-        />
-      </div>
-
-      <div className="text-center space-y-1 absolute bottom-4 w-full">
-        <div className="text-text-primary text24bold">
-          {percentage?.toFixed(2)}
-        </div>
-        <div className="text14medium text-text-tertiary">QFpower</div>
-      </div>
     </div>
   );
 }
